Add files via upload
I really hope that this fits into the main code of the game as seamlessly as possible. I tried to make it work and looked back at the main game code and tried to follow what was done in there. The green rectangle made at the start of the draw function was just a placeholder and can be removed/ignored.